Refactor State Pattern's code.
