+++ DeignPatternImplementationApp/DeignPatternImplementationApp/GoF Design Patterns/WorriorState.swift
@@ -1,0 +1,12 @@
+//
+//  State.swift
+//  DeignPatternImplementationApp
+//
+//  Created by Aaron Yen on 2022/8/26.
+//
+
+import Foundation
+
+protocol WarriorState {
+    func move(warrior: StatePattern.Warrior)
+}
